- [mod_userrepair] nofollow 連結，避免多重維護
git-svn-id: http://svn.openfoundry.org/pixmicat/Modules@711 2a703577-9a0d-0410-ac02-84028a1de3c1

+++ mod_userrepair/mod_userrepair.php
@@ -9,23 +9,35 @@ class mod_userrepair {
 	}
 
 	function autoHookThreadFront(&$txt,$isReply){
-		if(!$isReply) $txt.='<div style="text-align: right">[<a href="'.$this->SELF.'">討論串不見了？按一下這裡吧。</a>]</div>';
-		else $txt.='<div style="text-align: right">[<a href="'.$this->SELF.'&amp;res='.$isReply.'">文章不見了？按一下這裡吧。</a>]</div>';
+		if(!$isReply) $txt.='<div style="text-align: right">[<a href="'.$this->SELF.'" rel="nofollow">討論串不見了？按一下這裡吧。</a>]</div>';
+		else $txt.='<div style="text-align: right">[<a href="'.$this->SELF.'&amp;res='.$isReply.'" rel="nofollow">文章不見了？按一下這裡吧。</a>]</div>';
 	}
 
 	/* 模組獨立頁面 */
 	function ModulePage(){
 		global $PIO;
 		if(!isset($_GET['res'])) {
-			$PIO->dbMaintanence('repair',$PIO->dbMaintanence('repair'));
-			updatelog(); // 重導向到靜態快取
-			header('HTTP/1.1 302 Moved Temporarily');
-			header('Location: '.fullURL().PHP_SELF2.'?'.time());
+			if(!file_exists('./.userrepair')||isset($_GET['force'])) {
+				touch('./.userrepair');
+				$PIO->dbMaintanence('repair',$PIO->dbMaintanence('repair'));
+				updatelog(); // 重導向到靜態快取
+				unlink('./.userrepair');
+				header('HTTP/1.1 302 Moved Temporarily');
+				header('Location: '.fullURL().PHP_SELF2.'?'.time());
+			} else {
+				error('已經有其他人在修復中。<p>[<a href="'.$this->SELF.'&amp;force=1">強制執行</a>]</p>');
+			}
 		} else {
-			$no=intval($_GET['res']);
-			deleteCache(array($no));
-			header('HTTP/1.1 302 Moved Temporarily');
-			header('Location: '.fullURL().PHP_SELF.'?res='.$no);
+			if(!file_exists('./.userrepair')||isset($_GET['force'])) {
+				touch('./.userrepair');
+				$no=intval($_GET['res']);
+				deleteCache(array($no));
+				unlink('./.userrepair');
+				header('HTTP/1.1 302 Moved Temporarily');
+				header('Location: '.fullURL().PHP_SELF.'?res='.$no);
+			} else {
+				error('已經有其他人在修復中。<p>[<a href="'.$this->SELF.'&amp;res='.$_GET['res'].'&amp;force=1">強制執行</a>]</p>');
+			}
 		}
 	}
 
@@ -36,7 +48,7 @@ class mod_userrepair {
 
 	/* Get the module version infomation */
 	function getModuleVersionInfo(){
-		return 'Pixmicat! User Repair Module v080416';
+		return 'Pixmicat! User Repair Module v090526';
 	}
 
 }
